UBUNTU: add support for building selected stages of kernel
BugLink: http://bugs.launchpad.net/bugs/603087

To bootstrap in a cross-compile environment we need to be able to build
the compiler and libc, for this we need linux-libc-dev from the kernel.
However we cannot build the kernel without the compiler.  Catch 22.

This patch adds "DEB_STAGE=stage1" variable which skips building kernel
requesting only the linux-libc-dev package.

Based on a patch by Marcin Juszkiewicz <marcin.juszkiewicz@linaro.org>

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>