UBUNTU: Ubuntu-2.6.35-19.27
There is no real change here from 2.6.35-19.26.  Some udebs were
accidentally removed from the archive and thus the archive admin team
tried to upload a 2.6.35-19.27 kernel but didn't know to bump the ABI
which resulted in a build failure.  This just fixes up our repo and
changelog to match the version of the kernel uploaded to the archive.

Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>