HID: magicmouse: simplify touch data bit manipulation
The new format should be easier to read to determine which bits
correspond to which data. It also brings all the manipulation logic to
the top of the function. This makes size and orientation reading more
clear.

Note that the impetus for this change is the forthcoming support for the
Magic Trackpad, which has a different touch data protocol.

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Acked-by: Michael Poole <mdpoole@troilus.org>
Acked-by: Henrik Rydberg <rydberg@euromail.se>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>