DSS2: Don't power off a panel twice
BugLink: http://bugs.launchpad.net/bugs/588243

If we blank the panel by
echo 1 > /sys/devices/platform/omapfb/graphics/fb0/blank

Then, we reboot the sytem, the kernel will crash at
drivers/video/omap2/dss/core.c:323

This is because the panel is closed twice. Now check the state of a dssdev
to forbid a panel is power on or power off twice.

This has been submitted to upstream to the linux-omap mailing list:
http://marc.info/?l=linux-omap&m=128348306405397&w=2

Signed-off-by: Stanley.Miao <stanley.miao@windriver.com>
Signed-off-by: Bryan Wu <bryan.wu@canonical.com>
Signed-off-by: Mathieu Poirier <mathieu.poirier@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>