UBUNTU: SAUCE: hid: ntrig: identify firmware version (wiggled)
This adds firmware version polling to the end of probe and reports the
version both in the raw form and proccessed to match the formatting used
by N-Trig.

Has been tested succesfully on Dell Studio 17, Dell XT2, HP TX2 and
Lenovo T410s. Not likely to go upstream in its present form. However,
it does help constitue a tuning-free driver which works well together
with Unity, and as such are vital to the MT push.

Signed-off-by: Rafi Rubin <rafi@seas.upenn.edu>
Acked-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>
Acked-by: Jiri Slaby <jslaby@suse.cz>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Henrik Rydberg <rydberg@euromail.se>
Acked-by: Chase Douglas <chase.douglas@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>