xen, compat: Retruncate rax after ia32 syscall entry tracing
CVE-2010-3301

This is the counterpart for the upstream fix which was made as:

x86-64, compat: Retruncate rax after ia32 syscall entry tracing

In commit d4d6715, we reopened an old hole for a 64-bit ptracer touching a
32-bit tracee in system call entry.  A %rax value set via ptrace at the
entry tracing stop gets used whole as a 32-bit syscall number, while we
only check the low 32 bits for validity.

Fix it by truncating %rax back to 32 bits after syscall_trace_enter,
in addition to testing the full 64 bits as has already been added.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>