Revert "UBUNTU: SAUCE: blkfront: default to sd devices"
This reverts commit 38096c28f13d0c2dd08584ff834da6d81306c7b3. When using
PV on HVM the emulated devices can in some cases stay present and then
lead to a kernel bug because both the real and the virtual driver try
to use the sd* namespace.

Originally this patch was needed because EC2 passed sd* names to the
instance for the root device. Since we now use a label instead of the
block device name we can drop this patch and be more upstream compliant.

BugLink: http://bug.launchpad.net/bugs/684875

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>