Input: sysrq -- ensure sysrq_enabled and __sysrq_enabled are consistent
Currently sysrq_enabled and __sysrq_enabled are initialised separatly
and inconsitantly, leading to sysrq being actually enabled by reported
as not enabled in sysfs.  The first change to the sysfs configurable
synchronises these two:

    static int __read_mostly sysrq_enabled = 1;
    static int __sysrq_enabled;

Add a new configuration value which sets the default for these preventing
them becoming out of sync again.  Default this to 1 to mirror previous
behaviour.

Signed-off-by: Andy Whitcroft <apw@canonical.com>