SRU: xen: events: do not unmask event channels on resume
BugLink: http://bugs.launchpad.net/bugs/681083

SRU Justification:

Impact: With the current ec2 kernels the kernel oops described below is
experienced as a result of enabling interrupts on the pv spinlock event
channel.

Fix: The following patch is taken from upstream and is included in 2.6.37.
It has been reported to successfully prevent the oops.

Testcase: Migration of a guest (using suspend)

Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Brad Figg <brad.figg@canonical.com>